feat: add fuse.py
